fix imports after reorganizing demo folders
