+++ packages/web/demo/grid/index.html
@@ -16,6 +16,6 @@
     </style>
   </head>
   <body>
-    <script type="module" src="./grid-client.ts"></script>
+    <script type="module" src="./client.ts"></script>
   </body>
 </html>
